+++ INCommonsExample/INCommonsExample/HomeView.swift
@@ -18,7 +18,8 @@ struct HomeView: View {
 		ExampleViewFactory(title: "Share Sheet", view: { AnyView(ShareSheet()) }),
 		ExampleViewFactory(title: "View Condition", view: { AnyView(ViewConditions()) }),
 		ExampleViewFactory(title: "View Frame", view: { AnyView(ViewFrame()) }),
-		ExampleViewFactory(title: "Color Picker", view: { AnyView(ColorPickerExample()) })
+		ExampleViewFactory(title: "Color Picker", view: { AnyView(ColorPickerExample()) }),
+		ExampleViewFactory(title: "Overlay Window Presenter", view: { AnyView(OverlayWindowPresenterExample()) })
 	]
 
 	var body: some View {