+++ Sources/INCommons/UIKit/UIApplication+dismissKeyboard.swift
@@ -1,0 +1,9 @@
+import Foundation
+import UIKit
+
+public extension UIApplication {
+	/// Dismisses the keyboard.
+	func dismissKeyboard() {
+		sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
+	}
+}